[Flare] Move all event responders to dom directory (#15981)
