+++ packages/react-events/focus.js
@@ -9,6 +9,6 @@
 
 'use strict';
 
-const Focus = require('./src/Focus');
+const Focus = require('./src/dom/Focus');
 
 module.exports = Focus.default || Focus;
