+++ packages/react-events/src/dom/Swipe.js
